mcp/pyproject.toml
[project]
name = "mcptools"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.10.0",
    "uvicorn",
    "requests>=2.33.1",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "tavily-python>=0.7.23",
    "notion-client>=3.0.0",
    "python-dotenv>=1.2.2",
    "pandas>=3.0.2",
]
